Remove deprecated backend modules and refactor main application structure
- Deleted unused modules including auth, calendar, day, db, goal, http_handler, list, middleware, network, query, routine, and traversal to streamline the codebase.
- Updated the main application entry point to reflect the new module structure, focusing on AI and tool functionalities.
- Improved overall organization and maintainability of the backend code by consolidating relevant components.
