+++ backend/src/ai/mod.rs
@@ -1,0 +1,2 @@
+pub mod query;
+pub mod tool_registry;
