+++ backend/src/server/auth.rs
