+++ backend/src/server/mod.rs
@@ -1,0 +1,5 @@
+pub mod auth;
+pub mod db;
+pub mod http_handler;
+pub mod main;
+pub mod middleware;
