+++ backend/src/tools/day.rs
@@ -3,7 +3,7 @@ use chrono::{TimeZone, Utc};
 use neo4rs::{query, Graph};
 use serde_json::Value;
 
-use crate::goal::GOAL_RETURN_QUERY;
+use crate::tools::goal::GOAL_RETURN_QUERY;
 
 // Business logic functions with regular parameters
 pub async fn get_day_tasks(
